Create README.md
It is a solar grass cutter the main motive of this product is using of abundant sun light solar energy for agriculture purpose. It is works by using solar energy as the source of power. Where cutting blade is inserted at the down of the equipment where it rotates by motor using solar energy and moving wheels are present to move and it is controlled by wifi by using node mcu.